ActivitySpace.jl Documentation

ActivitySpace.jl Documentation

ActivitySpace.jl Documentation

Functions

Index

Functions

ActivitySpace.check_bias — Method.

check_bias(D; group_column::Symbol, group_a, group_b, X_column::Symbol, Y_column::Symbol, nreps::Int=500, sample_size::Int=100, pop_STP::Union{Number, Nothing}=nothing, f::Function=negative_exponential)

Returns a Float containing the difference between the population STP and the mean of the empirical sampling distribution for the given sample size.

source

ActivitySpace.city_sim_data — Function.

city_sim_data(city::AbstractString="")

Returns a dataset from a remote location.

Parameters

city - The name of the city for which you want simulated data. If left blank or if does not match an available city, the function will print a list of available city names.

Examples

Check available datasets:

julia> D  = city_sim_data()

Load the Utica simulation dataset:

julia> D  = city_sim_data("Utica")

Load the Buffalo simulation dataset:

julia> D  = city_sim_data("Buffalo")

source

ActivitySpace.data_prep — Method.

data_prep(D)

source

ActivitySpace.dataset — Method.

dataset(dataset_name::AbstractString)

Returns a dataset from the package's data directory.

Parameters

dataset_name - The name of the dataset (excluding file extension)

Examples

Load the Utica simulation dataset:

julia> D  = dataset("utica_sim0")

source

ActivitySpace.empirical_sampling_distribution — Method.

empirical_sampling_distribution(D, nreps, sample_size)

Returns a DataFrame containing the empirical sampling distribution for the Spatio-Temporal Proximity Index and its components.

source

ActivitySpace.randomize_race — Method.

randomize_race(D, rand_size)

Returns a DataFrame in which rand_size elements of the race column have been shuffled.

source

ActivitySpace.stprox — Method.

stprox(A::Array{Float64, 2}, B::Array{Float64, 2}; N_a=nothing, N_b=nothing, f::Function=negative_exponential)

Returns the Spatio-Temporal Proximity Index

source

ActivitySpace.stprox — Method.

stprox(D::DataFrame; group_column::Symbol, group_a, group_b, X_column::Symbol, Y_column::Symbol, N_a::Int=size(A, 1), N_b::Int=size(B, 1), f::Function=negative_exponential)

Returns the Spatio-Temporal Proximity Index

source

ActivitySpace.distance_sum — Function.

distance_sum(A::Array{<:Number, 2}, B::Union{Array{<:Number, 2}, Nothing}=nothing; f::Function=negative_exponential)::Float64

source

ActivitySpace.identity — Method.

identity(x::Number)::Float64

Internal method for calculating identity function. Since this is intended for use in distance calculations, $x$ must be a positive, single value. The function simply returns $x$.

source

ActivitySpace.negative_exponential — Method.

negative_exponential(x::Number)::Float64

Internal method for calculating negative exponential function. Since this is intended for use in distance calculations, $x$ must be a positive, single value. The function returns $e^{-x}$.

source

ActivitySpace.simple_distance — Method.

simple_distance(x1::Number, x2::Number, y1::Number, y2::Number, f::Function)::Float64

Internal method for calculating Euclidean distance between two points defined by $x1$, $y1$, and $x2$, $y2$. An optional function $f$ can be added to transform the distance result.

source

Index

ActivitySpace.check_bias

ActivitySpace.city_sim_data

ActivitySpace.data_prep

ActivitySpace.dataset

ActivitySpace.distance_sum

ActivitySpace.empirical_sampling_distribution

ActivitySpace.identity

ActivitySpace.negative_exponential

ActivitySpace.randomize_race

ActivitySpace.simple_distance

ActivitySpace.stprox

ActivitySpace.stprox