ActivitySpace.jl Documentation

ActivitySpace.jl Documentation

ActivitySpace.jl Documentation

Functions

Index

Functions

ActivitySpace.check_bias — Method.

check_bias(D;
    group_column::Symbol,
    group_a,
    group_b,
    X_column::Symbol,
    Y_column::Symbol,
    time_column::Symbol,
    ID_column::Symbol=:ID,
    nreps::Int=500,
    sample_size::Int=100,
    pop_STP::Union{Number, Nothing}=nothing,
    f::Function=negative_exponential)

Convenience function that calculates the difference between the mean of the empirical sampling distribution and the population STP value. Returns this difference (the estimator bias), along with the STP value, the full empirical sampling distribution, and information about the data. If the STP value is supplied in the function call, then it will not be calculated (thus saving processing time).

source

ActivitySpace.city_sim_data — Function.

city_sim_data(city::AbstractString="")

Returns a dataset from a remote location.

Parameters

city - The name of the city for which you want simulated data. If left blank or if does not match an available city, the function will print a list of available city names.

Examples

Check available datasets:

julia> D  = city_sim_data()

Load the Utica simulation dataset:

julia> D  = city_sim_data("Utica")

Load the Buffalo simulation dataset:

julia> D  = city_sim_data("Buffalo")

source

ActivitySpace.dataset — Method.

dataset(dataset_name::AbstractString)

Returns a dataset from the package's data directory.

Parameters

dataset_name - The name of the dataset (excluding file extension)

Examples

Load the Utica simulation dataset:

julia> D  = dataset("utica_sim0")

source

ActivitySpace.empirical_sampling_distribution — Method.

empirical_sampling_distribution(D::DataFrame;
    group_column::Symbol,
    group_a,
    group_b,
    X_column::Symbol,
    Y_column::Symbol,
    time_column::Symbol=:time,
    ID_column::Symbol=:ID,
    nreps::Int=500,
    sample_size::Int=100,
    f::Function=negative_exponential)::DataFrame

Returns a DataFrame containing the empirical sampling distribution for the Spatio-Temporal Proximity Index and its components.

source

ActivitySpace.negative_exponential — Method.

negative_exponential(x::Float64)::Float64

Returns $e^{-x}$.

source

ActivitySpace.randomize_column — Method.

randomize_column(D::DataFrame;
    column_name::Symbol,
    n::Int)

Returns a DataFrame in which n elements of the indicated column have been shuffled.

source

ActivitySpace.stprox — Method.

stprox(D::DataFrame;
    group_column::Symbol,
    group_a, group_b,
    X_column::Symbol,
    Y_column::Symbol,
    time_column::Symbol,
    N_a::Union{Int, Nothing}=nothing,
    N_b::Union{Int, Nothing}=nothing,
    f::Function=negative_exponential,
    time_approach=1)

Returns the Spatio-Temporal Proximity Index

source

ActivitySpace.distance_sum — Function.

distance_sum(A::Array{Float64, 2},
    B::Array{Float64, 2},
    f::Function=negative_exponential)::Float64

Internal function for summing distances between two sets of points, each represented in a matrix with 2 columns (X and Y coordinates). The distances are transformed by whatever function is supplied in the call.

source

ActivitySpace.distance_sum — Function.

distance_sum(A::Array{Float64, 2},
    f::Function=negative_exponential)::Float64

Internal function for summing distances between all points represented in a matrix with 2 columns containing the X and Y coordinates. The distances are transformed by whatever function is supplied in the call.

source

ActivitySpace.distance_sum3d — Function.

distance_sum3d(A::Array{Float64, 2}, B::Array{Float64, 2}, f::Function=negative_exponential)::Float64

Internal function for summing distances between two sets of points, each represented in a matrix with 3 columns (X and Y coordinates and time). The distances are transformed by whatever function is supplied in the call.

source

ActivitySpace.distance_sum3d — Function.

distance_sum3d(A::Array{Float64, 2}, f::Function=negative_exponential)::Float64

Internal function for summing distances between all points represented in a matrix with 3 columns containing the X and Y coordinates, and time. The distances are transformed by whatever function is supplied in the call.

source

Index

ActivitySpace.check_bias

ActivitySpace.city_sim_data

ActivitySpace.dataset

ActivitySpace.distance_sum

ActivitySpace.distance_sum

ActivitySpace.distance_sum3d

ActivitySpace.distance_sum3d

ActivitySpace.empirical_sampling_distribution

ActivitySpace.negative_exponential

ActivitySpace.randomize_column

ActivitySpace.stprox