ActivitySpace.jl Documentation

ActivitySpace.jl Documentation

ActivitySpace.jl Documentation

Functions

Index

Functions

ActivitySpace.calculate_bias — Method.

calculate_bias(D, randomns)

Returns a XXX.

source

ActivitySpace.check_bias — Method.

check_bias(pop_STP, D, nreps, sample_size)

Returns a Float containing the difference between the population STP and the mean of the empirical sampling distribution for the given sample size.

source

ActivitySpace.data_prep — Method.

data_prep(D)

source

ActivitySpace.dataset — Method.

dataset(dataset_name::AbstractString)

Returns a dataset from the package's data directory.

Parameters

dataset_name - The name of the dataset (excluding file extension)

Examples

Load the Utica simulation dataset:

julia> D  = dataset("utica_sim0")

source

ActivitySpace.empirical_sampling_distribution — Method.

empirical_sampling_distribution(D, nreps, sample_size)

Returns a DataFrame containing the empirical sampling distribution for the Spatio-Temporal Proximity Index and its components.

source

ActivitySpace.randomize_race — Method.

randomize_race(D, rand_size)

Returns a DataFrame in which rand_size elements of the race column have been shuffled.

source

ActivitySpace.stprox — Method.

stprox(A::Array{Float64, 2}, B::Array{Float64, 2}; N_a=nothing, N_b=nothing, f::Function=negative_exponential)

Returns the Spatio-Temporal Proximity Index

source

ActivitySpace.distance_sum — Function.

distance_sum(A::Array{<:Number, 2}, B::Union{Array{<:Number, 2}, Nothing}=nothing; f::Function=negative_exponential)::Float64

source

ActivitySpace.identity — Method.

identity(x::Number)::Float64

Internal method for calculating identity function. Since this is intended for use in distance calculations, $x$ must be a positive, single value. The function simply returns $x$.

source

ActivitySpace.negative_exponential — Method.

negative_exponential(x::Number)::Float64

Internal method for calculating negative exponential function. Since this is intended for use in distance calculations, $x$ must be a positive, single value. The function returns $e^{-x}$.

source

ActivitySpace.simple_distance — Method.

simple_distance(x1::Number, x2::Number, y1::Number, y2::Number, f::Function)::Float64

Internal method for calculating Euclidean distance between two points defined by $x1$, $y1$, and $x2$, $y2$. An optional function $f$ can be added to transform the distance result.

source

Index

ActivitySpace.calculate_bias

ActivitySpace.check_bias

ActivitySpace.data_prep

ActivitySpace.dataset

ActivitySpace.distance_sum

ActivitySpace.empirical_sampling_distribution

ActivitySpace.identity

ActivitySpace.negative_exponential

ActivitySpace.randomize_race

ActivitySpace.simple_distance

ActivitySpace.stprox